Feat: bubble 추가
