+++ src/components/molecules/whattodo/DepositSaving.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import Heart from '@/public/icons/heart.svg';
 import HeartClick from '@/public/icons/heartclick.svg';
-import Link from 'next/link';
 
 type TDepositSavingProps = {
   isLiked: boolean;
@@ -10,8 +9,6 @@ type TDepositSavingProps = {
   maxInterestRate: string;
   interestRate: string;
   onHeartClick: () => void;
-  productId: number;
-  productNum: number;
   size: 'Large' | 'Small';
 };
 
@@ -22,43 +19,37 @@ const DepositSaving: React.FC<TDepositSavingProps & React.HTMLAttributes<HTMLDiv
   maxInterestRate,
   interestRate,
   onHeartClick,
-  productId,
-  productNum,
   size,
   ...props
 }) => {
   return size === 'Large' ? (
-    <Link
-      href={{
-        pathname: `/whatToDo/${productId}`,
-        query: {
-          id: productId,
-          product: productNum,
-        },
-      }}
+    <div
+      {...props}
+      className='flex w-855 px-30 py-25 justify-between items-center gap-417 rounded-10 border-2 border-border01 bg-secondary cursor-pointer'
     >
-      <div
-        {...props}
-        className='flex w-855 px-30 py-25 justify-between items-center gap-417 rounded-10 border-2 border-border01 bg-secondary cursor-pointer'
-      >
-        <div className='flex flex-row items-center'>
-          <div className='w-72 h-72 rounded-13 bg-imageBase'></div>
-          <div className='ml-21'>
-            <div className='label-large text-typoPrimary'>{productName}</div>
-            <div className='mt-5 paragraph-large text-typoPrimary'>{bankName}</div>
-          </div>
-        </div>
-        <div className='flex flex-row items-center'>
-          <div className='flex flex-col text-right'>
-            <div className='label-large text-main'>최고 {maxInterestRate}%</div>
-            <div className='mt-5 paragraph-medium text-typoPrimary'>기본 {interestRate}%</div>
-          </div>
-          <button className='ml-19' onClick={onHeartClick}>
-            {isLiked ? <Heart className='w-37 h-37' /> : <HeartClick className='w-37 h-37' />}
-          </button>
+      <div className='flex flex-row items-center'>
+        <div className='w-72 h-72 rounded-13 bg-imageBase'></div>
+        <div className='ml-21'>
+          <div className='label-large text-typoPrimary'>{productName}</div>
+          <div className='mt-5 paragraph-large text-typoPrimary'>{bankName}</div>
         </div>
       </div>
-    </Link>
+      <div className='flex flex-row items-center'>
+        <div className='flex flex-col text-right'>
+          <div className='label-large text-main'>최고 {maxInterestRate}%</div>
+          <div className='mt-5 paragraph-medium text-typoPrimary'>기본 {interestRate}%</div>
+        </div>
+        <button
+          className='ml-19 z-toggle'
+          onClick={(event) => {
+            event.stopPropagation();
+            onHeartClick();
+          }}
+        >
+          {isLiked ? <Heart className='w-37 h-37' /> : <HeartClick className='w-37 h-37' />}
+        </button>
+      </div>
+    </div>
   ) : (
     <div
       {...props}
@@ -76,7 +67,13 @@ const DepositSaving: React.FC<TDepositSavingProps & React.HTMLAttributes<HTMLDiv
           <div className='label-medium text-main'>최고 {maxInterestRate}%</div>
           <div className='mt-5 paragraph-small text-typoPrimary'>기본 {interestRate}%</div>
         </div>
-        <button className='ml-10' onClick={onHeartClick}>
+        <button
+          className='ml-10'
+          onClick={(event) => {
+            event.stopPropagation();
+            onHeartClick();
+          }}
+        >
           {isLiked ? <Heart className='w-25 h-25' /> : <HeartClick className='w-25 h-25' />}
         </button>
       </div>
